# Erronka2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/talde2erronka2back/Erronka2/public/api/productos/ezabatu": {
            "put": {
                "tags": [
                    "Produktuak"
                ],
                "summary": "Produktuak ezabatzeko metodoa",
                "operationId": "499a2809455094f1f5bf7c3e642cff70",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Produktuaren IDa",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produktuen datuak ezabatu dira."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/productos/editatu": {
            "put": {
                "tags": [
                    "Produktuak"
                ],
                "summary": "Produktuak editateko metodoa",
                "operationId": "fce8593f247e5e2c1bd1b70060e1c1f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "izena",
                                    "marka",
                                    "kategoria",
                                    "deskribapena",
                                    "stock",
                                    "stock_alerta"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Produktuaren IDa",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "izena": {
                                        "description": "Produktuaren izena",
                                        "type": "string"
                                    },
                                    "marka": {
                                        "description": "Produktuaren marka",
                                        "type": "string"
                                    },
                                    "kategoria": {
                                        "description": "Produktuaren kategoria IDa",
                                        "type": "integer"
                                    },
                                    "deskribapena": {
                                        "description": "Produktuaren deskribapena",
                                        "type": "string"
                                    },
                                    "stock": {
                                        "description": "Produktuaren stock-eko kantitatea",
                                        "type": "integer"
                                    },
                                    "stock_alerta": {
                                        "description": "Stock alerta-ko kantitatea",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produktuen datuak editatu dira."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/productos/txertatu": {
            "post": {
                "tags": [
                    "Produktuak"
                ],
                "summary": "Produktuak txertatzeko metodoa",
                "operationId": "9e5a3b08eab14a3c987fff60163c4de0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "izena",
                                    "marka",
                                    "kategoria",
                                    "deskribapena",
                                    "stock",
                                    "stock_alerta"
                                ],
                                "properties": {
                                    "izena": {
                                        "description": "Produktuaren izena",
                                        "type": "string"
                                    },
                                    "marka": {
                                        "description": "Produktuaren marka",
                                        "type": "string"
                                    },
                                    "kategoria": {
                                        "description": "Produktuaren kategoria IDa",
                                        "type": "integer"
                                    },
                                    "deskribapena": {
                                        "description": "Produktuaren deskribapena",
                                        "type": "string"
                                    },
                                    "stock": {
                                        "description": "Produktuaren stock-eko kantitatea",
                                        "type": "integer"
                                    },
                                    "stock_alerta": {
                                        "description": "Stock alerta-ko kantitatea",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produktuen datuak txertatu dira."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/productos/erosi": {
            "post": {
                "tags": [
                    "Produktuak"
                ],
                "summary": "Produktuak erositzeko metodoa",
                "operationId": "b3080461d459a7a7d69fad5d473e1d75",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "produktua_id",
                                    "kantitatea",
                                    "ordu_kopurua"
                                ],
                                "properties": {
                                    "produktua_id": {
                                        "description": "Erosi nahi den produktuaren IDa",
                                        "type": "integer"
                                    },
                                    "kantitatea": {
                                        "description": "Erosi nahi den produktuaren kantitatea",
                                        "type": "integer"
                                    },
                                    "ordu_kopurua": {
                                        "description": "Erosketa ordu kopurua",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produktuen datuak txertatu dira."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}